+++ lib/config/__init__.py
@@ -1,6 +1,6 @@
 """The config core library."""
 
-from ._pydantic import PydanticYamlModel
+from .pydantic import PydanticYamlModel
 
 __all__ = [
     # Model.
