+++ lib/config/pydantic.py
