+++ lib/evagg/__init__.py
@@ -1,12 +1,12 @@
 """The evagg core library."""
 
-from ._app import SynchronousLocalApp
-from ._content._prompt_based import PromptBasedContentExtractor
-from ._content._simple import SimpleContentExtractor
-from ._content._truth_set import TruthsetContentExtractor
-from ._interfaces import IEvAggApp, IExtractFields, IGetPapers, IWriteOutput
-from ._io import ConsoleOutputWriter, FileOutputWriter, TableOutputWriter
-from ._library import PubMedFileLibrary, SimpleFileLibrary, TruthsetFileLibrary
+from .app import SynchronousLocalApp
+from ._content.prompt_based import PromptBasedContentExtractor
+from ._content.simple import SimpleContentExtractor
+from ._content.truth_set import TruthsetContentExtractor
+from .interfaces import IEvAggApp, IExtractFields, IGetPapers, IWriteOutput
+from .io import ConsoleOutputWriter, FileOutputWriter, TableOutputWriter
+from .library import PubMedFileLibrary, SimpleFileLibrary, TruthsetFileLibrary
 
 __all__ = [
     # Interfaces.
