+++ lib/evagg/lit/pubmed/__init__.py
@@ -1,6 +1,6 @@
 """Utilities for interacting with PubMed."""
 
-from ._annotation import PubtatorEntityAnnotator
+from .annotation import PubtatorEntityAnnotator
 
 __all__ = [
     # Annotation.
