+++ lib/evagg/types/__init__.py
@@ -1,8 +1,8 @@
 """Base types for the evagg library."""
 
-from ._base import Paper, Variant
-from ._interfaces import IPaperQuery, IPaperQueryIterator
-from ._query import Query, QueryIterator
+from .base import Paper, Variant
+from .interfaces import IPaperQuery, IPaperQueryIterator
+from .query import Query, QueryIterator
 
 __all__ = [
     # Base.
