+++ lib/evagg/types/base.py
