+++ lib/evagg/web/entrez/__init__.py
@@ -1,7 +1,7 @@
 """Classes for interacting with Entrez APIs."""
 
-from ._client import BioEntrezClient, BioEntrezConfig, BioEntrezDotEnvConfig
-from ._interfaces import IEntrezClient
+from .client import BioEntrezClient, BioEntrezConfig, BioEntrezDotEnvConfig
+from .interfaces import IEntrezClient
 
 __all__ = [
     # Interfaces.
