+++ lib/execute/__init__.py
@@ -3,6 +3,6 @@
 Takes as input a yaml configuration file and runs the evidence aggregation pipeline.
 """
 
-from ._cli import run_sync
+from .cli import run_sync
 
 __all__ = ["run_sync"]
