+++ test/evagg/test_library.py
@@ -95,7 +95,7 @@ def test_pubmedfilelibrary(mock_build_papers, mock_find_pmids_for_gene, entrez_c
 # Test if an incorrect gene name is being passed in. If so, return an empty list.
 def test_find_pmids_for_gene_invalid_gene(mocker, entrez_client):
     # Mock the esearch method to return an XML string without an "IdList" element
-    mock_esearch = mocker.patch("lib.evagg._library.IEntrezClient.esearch", return_value="<root></root>")
+    mock_esearch = mocker.patch("lib.evagg.library.IEntrezClient.esearch", return_value="<root></root>")
 
     # PubMedFileLibrary instance
     library = PubMedFileLibrary(entrez_client, max_papers=1)
@@ -114,7 +114,7 @@ def test_find_pmids_for_gene_invalid_gene(mocker, entrez_client):
 def test_find_pmids_for_gene_no_papers(mocker, entrez_client):
     # Mock the esearch method to return an XML string without any "Id" elements
     mock_esearch = mocker.patch(
-        "lib.evagg._library.IEntrezClient.esearch", return_value="<root><IdList></IdList></root>"
+        "lib.evagg.library.IEntrezClient.esearch", return_value="<root><IdList></IdList></root>"
     )
 
     # PubMedFileLibrary instance
